Move shared tests to the correct directory
