+++ frontend/src/embedding-sdk-bundle/components/public/dashboard/EditableDashboard/EditableDashboard.unit.spec.tsx
@@ -10,7 +10,7 @@ import { screen, waitFor, within } from "__support__/ui";
 import { ROOT_COLLECTION } from "metabase/entities/collections";
 import { createMockCollection } from "metabase-types/api/mocks";
 
-import { addSubscriptionTests } from "../StaticDashboard/sharedTests.spec";
+import { addSubscriptionTests } from "../sharedTests.spec";
 import {
   type SetupSdkDashboardOptions,
   setupSdkDashboard,
